Backend/requirements.txt
fastapi
uvicorn
requests
Pillow
numpy
opencv-python